Bubble up some functionality from Tag to TagWithTooltip and TagDeleteBadge
